fix compiler warning
```
tests\test_read.c(85): warning C4018: '<': signed/unsigned mismatch
```

+++ tests/test_read.c
@@ -82,7 +82,7 @@ int test(LIBSSH2_SESSION *session)
             fprintf(stderr, "Unable to read response: %d\n", (int)err);
         else {
             unsigned int i;
-            for(i = 0; i < err; ++i) {
+            for(i = 0; i < (unsigned long)err; ++i) {
                 if(buf[i]) {
                     fprintf(stderr, "Bad data received\n");
                     /* Test will fail below due to bad data length */
